replace cst json with structs
