+++ edb/edgeql-parser/src/lib.rs
@@ -1,11 +1,11 @@
 pub mod ast;
-pub mod cparser;
 pub mod expr;
 pub mod hash;
 pub mod helpers;
 #[cfg(feature = "python")]
 pub mod into_python;
 pub mod keywords;
+pub mod parser;
 pub mod position;
 pub mod preparser;
 pub mod schema_file;
